+++ tests/installtest/test_install.sh
@@ -471,22 +471,6 @@ function test1()
   cleanup_egg_file
 }
 
-# Install test via setup.py has been disabled
-function test2()
-{
-  testcase="test2"
-  info "Testcase $testcase: setup.py install from repo root directory: $ROOT_DIR (deprecated by setuptools)"
-  make_virtualenv "$testcase"
-
-  call "cd $ROOT_DIR; python setup.py install" "Installing with setup.py from repo root directory (latest package levels)"
-
-  assert_import_ok "pywbem"
-  assert_import_ok "pywbem_mock"
-  assert_no_moftab_build
-  remove_virtualenv "$testcase"
-  cleanup_egg_file
-}
-
 function test3()
 {
   testcase="test3"
@@ -509,23 +493,6 @@ function test4()
   make_virtualenv "$testcase"
 
   call "cd $TMP_TEST_DIR; pip install $(abspath $SRC_DISTFILE) $PIP_OPTS" "Installing with pip from source distribution archive (PACKAGE_LEVEL=$PACKAGE_LEVEL)"
-
-  assert_import_ok "pywbem"
-  assert_import_ok "pywbem_mock"
-  assert_no_moftab_build
-  remove_virtualenv "$testcase"
-  cleanup_egg_file
-}
-
-# Install test via setup.py has been disabled
-function test5()
-{
-  testcase="test5"
-  info "Testcase $testcase: setup.py install from unpacked source distribution archive: $SRC_DISTFILE (deprecated by setuptools)"
-  make_virtualenv "$testcase"
-  run "tar -x -v -f $SRC_DISTFILE -C $SRC_DISTFILE_UNPACK_DIR" "Unpacking source distribution archive to: $SRC_DISTFILE_UNPACK_DIR"
-
-  call "cd $SRC_DISTFILE_UNPACK_DIR/$SRC_DISTFILE_TOP_DIR; python setup.py install" "Installing with setup.py from unpack directory: $SRC_DISTFILE_UNPACK_DIR/$SRC_DISTFILE_TOP_DIR (latest package levels)"
 
   assert_import_ok "pywbem"
   assert_import_ok "pywbem_mock"